restructure update student user view
